requirements.txt
# ============================================
# Dependencias del Proyecto EcoTech Solutions
# ============================================
# Instalar todas las dependencias con:
# pip install -r requirements.txt
# ============================================

# Conexión a MySQL
mysql-connector-python==8.2.0

# Hasheo seguro de contraseñas
bcrypt==4.1.2

# Peticiones HTTP para API
requests==2.31.0

# Carga de variables de entorno
python-dotenv==1.0.0

# ============================================
# INSTALACIÓN RÁPIDA:
# ============================================
# 1. Abre una terminal en la carpeta del proyecto
# 2. Ejecuta: pip install -r requirements.txt
# 3. Espera a que se instalen todas las librerías
# 4. ¡Listo para ejecutar!
# ============================================
